+++ logging/CHANGELOG.md
@@ -10,5 +10,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 - Move logging logic to shared crate
 
+### Changed
+- Default log level for debug builds is debug
+
 
 [Unreleased]: https://github.com/replicante-io/common/compare/v0.1.0...HEAD